chore: move objectToMultihash into func
